requirements.txt
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
python-dotenv
